Fix receive method
This fix ensures that the server and microscope has time to finish a
command and return a reply, before method receive returns. Some
commands, like change of PMT gain, can take a while to excecute for the
microscope. Start scan command also takes a while before the server
returns a reply.

* Add method _check_socket to check if socket is readable using select.
* Use _check_socket in method receive until socket is readable or after
  timeout of 10 seconds. When socket is readable, call recv on socket.
* Fix PEP issues in cam module.
* Fix docstrings in cam module.
* Add method fileno to mock socket class, in test_cam module, and make
  select work during tests.
* Fix some, but not all, PEP issues in test_cam module.